+++ inc/customizer/sections/section-typography.php
@@ -20,8 +20,8 @@ function mh_register_typography_section( $wp_customize ) {
 	// ==============================================
 	$wp_customize->add_section( 'magic_hat_typography', array(
 		'title'    => __( '🪶 Typography', 'xophz-magic-hat' ),
-		'priority' => 20,
-		'panel'    => 'magic_hat_site_styles',
+		'priority' => 50,
+		'panel'    => 'magic_hat_general_settings',
 	) );
 
 	// Base Font Family